+++ src/index.ts
@@ -1,6 +1,5 @@
 export * from "./AuthContext";
 export * from "./AuthProvider";
-export type { AuthProviderProps } from "./AuthProps";
 export type { AuthState } from "./AuthState";
 export * from "./useAuth";
 export { hasAuthParams } from "./utils";
